Move and rename write_subtitles for better organization
